requirements.txt
numpy>=1.16.4,<1.19.0
jax>=0.2.6
jaxlib>=0.1.57
flax>=0.2.2
opencv-python>=4.4.0
Pillow>=7.2.0
pyyaml>=5.3.1
tensorboard>=2.4.0
tensorflow>=2.3.1
tensorflow-hub>=0.11.0